requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   -r requirements.in
    #   google-genai
    #   httpx
    #   openai
    #   sse-starlette
    #   starlette
    #   watchfiles
arq==0.26.3
    # via
    #   -r requirements.in
    #   sentry-sdk
async-timeout==5.0.1
    # via -r requirements.in
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
bcrypt==4.3.0
    # via chromadb
build==1.3.0
    # via chromadb
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
chromadb==1.0.21
    # via -r requirements.in
click==8.2.1
    # via
    #   arq
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
coverage==7.10.6
    # via pytest-cov
distro==1.9.0
    # via
    #   openai
    #   posthog
durationpy==0.10
    # via kubernetes
fastapi==0.116.1
    # via
    #   -r requirements.in
    #   sentry-sdk
filelock==3.19.1
    # via huggingface-hub
flatbuffers==25.2.10
    # via onnxruntime
fsspec==2025.9.0
    # via huggingface-hub
google-api-core==2.25.1
    # via -r requirements.in
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-genai
    #   kubernetes
google-genai==1.36.0
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio==1.74.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.9
    # via huggingface-hub
hiredis==3.2.1
    # via redis
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   chromadb
    #   google-genai
    #   openai
huggingface-hub==0.34.4
    # via tokenizers
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
importlib-resources==6.5.2
    # via chromadb
iniconfig==2.1.0
    # via pytest
jiter==0.10.0
    # via openai
jsonschema==4.25.1
    # via chromadb
jsonschema-specifications==2025.9.1
    # via jsonschema
kubernetes==33.1.0
    # via chromadb
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via chromadb
mpmath==1.3.0
    # via sympy
numpy==2.3.3
    # via
    #   chromadb
    #   onnxruntime
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
onnxruntime==1.22.1
    # via chromadb
openai==1.107.1
    # via -r requirements.in
opentelemetry-api==1.37.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.37.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.37.0
    # via chromadb
opentelemetry-proto==1.37.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.37.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.58b0
    # via opentelemetry-sdk
orjson==3.11.3
    # via chromadb
overrides==7.7.0
    # via chromadb
packaging==25.0
    # via
    #   build
    #   huggingface-hub
    #   onnxruntime
    #   pytest
pillow==11.3.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
posthog==5.4.0
    # via chromadb
proto-plus==1.26.1
    # via google-api-core
protobuf==6.32.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pybase64==1.4.2
    # via chromadb
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   chromadb
    #   fastapi
    #   google-genai
    #   openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via redis
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.2.0
    # via build
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
    #   pytest-snapshot
pytest-asyncio==1.2.0
    # via -r requirements.in
pytest-cov==7.0.0
    # via -r requirements.in
pytest-mock==3.15.1
    # via -r requirements.in
pytest-snapshot==0.9.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   posthog
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
python-json-logger==3.3.0
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   uvicorn
redis==5.3.1
    # via
    #   -r requirements.in
    #   arq
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-genai
    #   huggingface-hub
    #   kubernetes
    #   posthog
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rich==14.1.0
    # via
    #   chromadb
    #   typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
sentry-sdk==2.37.1
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   kubernetes
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sse-starlette==3.0.2
    # via -r requirements.in
sseclient-py==1.8.0
    # via -r requirements.in
starlette==0.47.3
    # via fastapi
sympy==1.14.0
    # via onnxruntime
tenacity==9.1.2
    # via
    #   chromadb
    #   google-genai
tokenizers==0.22.0
    # via chromadb
tqdm==4.67.1
    # via
    #   chromadb
    #   huggingface-hub
    #   openai
typer==0.17.4
    # via chromadb
typing-extensions==4.15.0
    # via
    #   chromadb
    #   fastapi
    #   google-genai
    #   huggingface-hub
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   kubernetes
    #   requests
    #   sentry-sdk
uvicorn==0.35.0
    # via
    #   -r requirements.in
    #   chromadb
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websocket-client==1.8.0
    # via kubernetes
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
zipp==3.23.0
    # via importlib-metadata
